# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital Payments API",
        "description": "API documentation for the Digital Payments system",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/wallets/balance": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Get user wallet balance",
                "description": "Returns the current balance of the user's wallet. This is a read-only operation (Query) and does not modify any system state.",
                "operationId": "getWalletBalance",
                "responses": {
                    "200": {
                        "description": "Wallet balance retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "string",
                                            "example": "c1a9b2e4-7f12-4a88-9c01-9bdeff123456"
                                        },
                                        "balance": {
                                            "description": "Available wallet balance",
                                            "type": "number",
                                            "format": "float",
                                            "example": 150.75
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wallet not found for the given user"
                    },
                    "401": {
                        "description": "Unauthenticated user"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}